chore: Update Nav component link to homepage and index page layout
Refactor the Nav component in Nav.jsx to fix the link to the homepage, updating it to point to "/index" instead of "/". This change ensures that the link correctly navigates to the index page of the application. Additionally, update the index page layout in index.jsx by adding a marginLeft style of '15%' to the main container div. This change improves the visual alignment of the page.